Force bash code, written and executed as part of a make target, to exit in case of an error
When running `pull-images`, it seems that the for-loop runs in a sub-shell so that
make does not notice the exit != 0, thus continues to run. This behaviour is not
intended, instead the process should stop to indicate to the user that sth went wrong.
This Will also be useful, when running in CI.